Create a modal for dictionary
